Move teardown code to a separate package
