+++ java/clients/src/main/java/sleeper/clients/teardown/RemoveJarsBucket.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package sleeper.clients.deploy;
+package sleeper.clients.teardown;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
